Move API calls to their own file
